refactor: update import path for stream handler schema and fix casing
